inventory 1.1
