+++ src/GUI/view/view/agentView/AgentView.java
@@ -15,10 +15,8 @@ public abstract class AgentView extends JButton implements View, ActionListener 
      * initializing the button
      */
     public AgentView() {
-        setSize(100, 50);
-        setContentAreaFilled(false);
-        setBorderPainted(false);
-        setFocusPainted(false);
+
+
     }
 
     public void setImageIcon(ImageIcon imageIcon) {
@@ -39,20 +37,5 @@ public abstract class AgentView extends JButton implements View, ActionListener 
     @Override
     public void onClick() {
 
-    }
-
-    @Override
-    /**
-     * Print a black rectangle with rounded corners. This method can be called from the paint methods of classes, who inherit from this class,
-     * when putting an image on top of it.
-     */
-    public void paint(Graphics g) {
-        setOpaque(false);
-        Graphics2D g2 = (Graphics2D) g.create();
-        g2.setColor(Color.BLACK);
-        g2.setRenderingHints(new RenderingHints(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON));
-        g2.fillRoundRect(0, 0, getWidth(), getHeight(), 40, 40);
-        g2.dispose();
-        super.paint(g);
     }
 }
