+++ src/GUI/view/view/agentView/EmptyAgentInventorySlotView.java
@@ -4,6 +4,6 @@ import javax.swing.*;
 
 public class EmptyAgentInventorySlotView extends AgentView {
     public EmptyAgentInventorySlotView() {
-        setBorder(BorderFactory.createEmptyBorder(35, 35, 35, 35));
+
     }
 }
